// src/api/openapi/models/UpdateDraftEvent.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UpdateDraftEvent = {
    event_start?: string;
    event_end?: string;
    location?: string;
    title: string;
    about?: string;
    type?: string;
    guide?: string;
    image?: Blob;
};